requirements.txt
Django
gunicorn
uvicorn
whitenoise
django-autocomplete-light
channels
daphne
django-extensions
djangorestframework
django-cleanup
django-cors-headers
django-filter
django-excel-response
xlwt
django-simple-history
django-soft-delete
twilio
openpyxl
Pillow